+++ sanitizer/src/test/java/com/networknt/sanitizer/SanitizerHandlerTest.java
@@ -20,6 +20,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.networknt.client.Http2Client;
 import com.networknt.config.Config;
 import com.networknt.exception.ClientException;
+import com.networknt.sanitizer.builder.ServerBuilder;
 import io.undertow.Undertow;
 import io.undertow.client.*;
 import io.undertow.util.*;
