+++ sanitizer/src/test/java/com/networknt/sanitizer/SanitizerHandlerWithEncodeTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.networknt.client.Http2Client;
 import com.networknt.config.Config;
 import com.networknt.exception.ClientException;
+import com.networknt.sanitizer.builder.ServerBuilder;
 import io.undertow.Undertow;
 import io.undertow.client.ClientConnection;
 import io.undertow.client.ClientRequest;
