+++ sanitizer/src/test/resources/config/sanitizer_with_encode_type.yml
@@ -10,4 +10,4 @@ sanitizeBody: true
 # if it is enabled, does header need to be sanitized
 sanitizeHeader: true
 
-encoding: 'example'
+encoding: 'fake'
